requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-vision==3.7.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.0
oauth2client==4.1.3
pillow==10.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
